Se creo un context para actividades
