+++ src/context/actividades/index.js
@@ -1,0 +1,5 @@
+import {createContext} from 'react'
+
+const ActividadesContext = createContext('actividades')
+
+export default ActividadesContext
