+++ src/pages/Actividades.js
@@ -1,15 +1,13 @@
-import React, { useState } from 'react'
+import React, { useContext, useEffect } from 'react'
 import './styles/actividades.css'
 import Cronograma from '../components/Cronograma'
 import FormMes from '../components/FormMes'
 import Formsemana from '../components/Formsemana'
-import { useEffect } from 'react/cjs/react.development'
 import Loader from '../components/Loader'
+import ActividadesContext from '../context/actividades'
 
 function Actividades () {
-    const [turnos, setTurnos] = useState([])
-    const [turnosJov, setTurnoJov] = useState([])
-    const [cargando, setCargando] = useState(true)
+    const {cargando, turnosJov, turnos, getDatos} = useContext(ActividadesContext)
 
     useEffect(() => {
         getDatos()
@@ -18,21 +16,6 @@ function Actividades () {
     useEffect(() => {
         //
     }, [turnos])
-
-    const getDatos = async () => {
-        setCargando(true)
-        try {
-            await fetch('https://uecapi.herokuapp.com/cronograma/getTurnoMensual.php')
-                .then(response => response.json())
-                .then(data => setTurnos(data))
-            await fetch('https://uecapi.herokuapp.com/cronograma/getTurnoJovenes.php')
-                .then(response => response.json())
-                .then(data => setTurnoJov(data))
-            setCargando(false)
-        } catch (error) {
-            console.log('Ocurrio un error al traer los datos')
-        }
-    }
 
     if (cargando) {
         return(
